adapt account page for single household
