+++ flutter_application/lib/SignInPage.dart
@@ -2,6 +2,7 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 import 'auth_service.dart'; // Import the AuthService
 import 'HomePage.dart';
+import 'user_model.dart';
 
 class SignInPage extends StatefulWidget {
   const SignInPage({super.key});
@@ -162,31 +163,5 @@ class _SignInPageState extends State<SignInPage> {
       ).toJson();
 
       userCollection.doc(id).set(newUser);
-  }
-}
-
-// User model: could be put in a different file in the future but for now is here
-
-class UserModel{
-  final String? email;
-  final String? password;
-  final String? id;
-
-  UserModel( this.id, this.email, this.password);
-
-  static UserModel fromSnapshot(DocumentSnapshot<Map<String, dynamic>> snapshot){
-    return UserModel(
-      snapshot['id'], 
-      snapshot['email'], 
-      snapshot['password']
-    );
-  }
-
-  Map<String, dynamic> toJson(){
-    return{
-      "id": id,
-      "email": email,
-      "password": password,
-    };
   }
 }
